refactor: move routes to constants
